Interactive camera, subpixel jitter, DoF, sky lighting
- New interactive camera, making the engine much more user-friendly. Basically a port of the one from my OpenGL project, GLEngine.
Was a bit tricky to use GLFW callbacks since it is written in C and do not accept a method that is a member of a class as the argument of glfwSetKeyCallback() for example.
So I went into the dirty trick road and used ImGui (which actually received information from GLFW callback functions) indirectly to reproduce the GLFW callback system. It is still kinda dirty of course, but it does the job and is enough for now.
- As a consequence of the addition of the camera, we now have subpixel jitter, which give us nearly free supersampling.
- Camera also means Depth of Field. It is a simple implementation, and the user can tweak the aperture radius and the focal distance in order to achieve the desired look.
- Added sky lighting, making the scene more alive, and also a lot faster to converge !
- New test scene, will have to do a more suitable Cornell scene later.